--- src/new-with-leveldb/lib.rs
@@ -1,3 +1,0 @@
-pub mod encoder;
-pub use encoder::{encode,decode,Decoded,DecodedNode,DecodedWay,DecodedRelation};
-mod varint;
